feat: rename icon-name.type.ts to icon-name.ts
there's no point in naming it icon-name.type, if it's already in the
"types" directory xD
